@@ -1,6 +1,6 @@
 {
   "name": "three",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -9,11 +9,12 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "ant-design-vue": "^1.6.2",
+    "ant-design-vue": "^2.0.0",
     "axios": "^0.19.2",
     "babel-runtime": "^6.26.0",
     "core-js": "^3.6.5",
     "full-icu": "^1.3.1",
+    "markdown-vue-next": "https://github.com/mengdaoshizhongxinyang/Markdown-Vue-Next.git",
     "mockjs2": "^1.0.8",
     "monaco-editor": "^0.20.0",
     "postprocessing": "^6.8.8",
@@ -22,26 +23,31 @@
     "three": "^0.110.0",
     "ts-loader": "^8.0.4",
     "typescript": "^4.0.3",
-    "vue": "^2.6.10",
-    "vue-draggable-resizable": "^2.1.0",
-    "vue-loader": "^15.7.1",
+    "vue": "^3.0.0",
+    "vue-loader": "^16.0.0",
     "vue-ls": "^3.2.1",
     "vue-markdown": "^2.2.4",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.2"
+    "vue-router": "^4.0.1",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.4.6",
+    "@types/vue-markdown": "^2.2.1",
+    "@types/webpack-env": "^1.15.3",
+    "@vue/babel-plugin-jsx": "^1.0.0-rc.3",
+    "@vue/cli-plugin-babel": "^4.5.8",
+    "@vue/cli-plugin-router": "^4.5.8",
     "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-service": "^2.0.0",
+    "@vue/cli-service": "^4.0.0",
+    "@vue/compiler-sfc": "^3.0.2",
     "@vue/test-utils": "^1.0.3",
     "babel-plugin-import": "^1.13.0",
-    "css-loader": "^3.6.0",
+    "css-loader": "^5.0.1",
     "html-webpack-plugin": "^4.5.0",
     "less": "^3.9.0",
     "less-loader": "^5.0.0",
     "monaco-editor-webpack-plugin": "^1.9.0",
-    "vue-markdown-loader": "^2.4.1",
-    "vue-template-compiler": "^2.6.10"
+    "typescript-plugin-css-modules": "^3.2.0",
+    "typings-for-css-modules-loader": "^1.7.0",
+    "vue-markdown-loader": "^2.4.1"
   }
 }
