@@ -1,21 +1,24 @@
-import Vue from 'vue'
-import App from './App.vue'
+import {createApp} from 'vue'
+import App from './App'
+// import Icon from '@ant-design/icons-vue';
 import router from './router'
 import './mock'
 import 'ant-design-vue/dist/antd.css' 
 import Directives from '@/utils/directives.js';
-import Storage from 'vue-ls';
- 
+import Storage from '@/utils/localstorage';
+import { Store } from "@/types/store";
 let options = {
   namespace: 'vuejs__', // key prefix
   name: 'ls', // name variable Vue.[ls] or this.[$ls],
   storage: 'local', // storage name session, local, memory
 };
-Vue.use(Storage, options);
-import store, { Store } from './store/index';
+const Vue=createApp(App).use(Storage,options).use(router);
 
+// Vue.use
+import store from './store/index';
+Vue.use(store)
 Vue.use(Directives)
-Vue.config.productionTip = false
+// Vue.config.productionTip = false
 import {
   LocaleProvider,
   Layout,
@@ -33,7 +36,7 @@ import {
   Modal,
   Table,
   Tabs,
-  Icon,
+  // Icon,
   Badge,
   Popover,
   Dropdown,
@@ -69,7 +72,7 @@ Vue.use(Layout)
 Vue.use(Carousel)
 
 Vue.use(Input)
-Vue.use(InputNumber)
+// Vue.use(InputNumber)
 Vue.use(Button)
 Vue.use(Switch)
 Vue.use(Radio)
@@ -82,7 +85,7 @@ Vue.use(Col)
 Vue.use(Modal)
 Vue.use(Table)
 Vue.use(Tabs)
-Vue.use(Icon)
+// Vue.use(Icon)
 Vue.use(Badge)
 Vue.use(Popover)
 Vue.use(Dropdown)
@@ -104,23 +107,22 @@ Vue.use(Progress)
 Vue.use(Skeleton)
 Vue.use(Popconfirm)
 // Vue.use(VueCropper)
+// app.config.globalProperties.$message = message;
+Vue.config.globalProperties.$confirm = Modal.confirm
+Vue.config.globalProperties.$message = message
+Vue.config.globalProperties.$notification = notification
+Vue.config.globalProperties.$info = Modal.info
+Vue.config.globalProperties.$success = Modal.success
+Vue.config.globalProperties.$error = Modal.error
+Vue.config.globalProperties.$warning = Modal.warning
+// const app=new Vue({
+//   router,
+//   store:store,
+//   render: h => h(App)
+// }).$mount('#app')
 
-Vue.prototype.$confirm = Modal.confirm
-Vue.prototype.$message = message
-Vue.prototype.$notification = notification
-Vue.prototype.$info = Modal.info
-Vue.prototype.$success = Modal.success
-Vue.prototype.$error = Modal.error
-Vue.prototype.$warning = Modal.warning
-const app=new Vue({
-  router,
-  store:store,
-  render: h => h(App)
-}).$mount('#app')
-
-Vue.prototype.$tstore = app.$store;
-app.$tstore.dispatch('initAll')
-Vue.prototype.$aemit = (that: any, mutationName: string, ...params: any) => {
+// Vue.config.globalProperties.$tstore = Vue.config.globalProperties.$store
+Vue.config.globalProperties.$aemit = (that: any, mutationName: string, ...params: any) => {
   if (!mutationName) throw new Error('$aemit need mutationName param');
   return new Promise((resolve, reject) => {
     that.$emit(mutationName, ...params, {
@@ -129,9 +131,11 @@ Vue.prototype.$aemit = (that: any, mutationName: string, ...params: any) => {
     });
   });
 };
-declare module 'vue/types/vue' {
-  interface Vue {
-    $tstore: Store;
-    $aemit:<T>(that: any, mutationName: string, ...params: any) => Promise<T>;
-  }
-}
+Vue.mount('#app')
+
+
+// declare module 'vue/types/vue' {
+//   interface Vue {
+//     $aemit:<T>(that: any, mutationName: string, ...params: any) => Promise<T>;
+//   }
+// }
