@@ -1,18 +1,9 @@
 module.exports = {
   presets: [
-    '@vue/app',
-    [
-      '@babel/preset-env',
-      {
-        'useBuiltIns': 'entry'
-      }
-    ]
+    '@vue/cli-plugin-babel/preset'
   ],
   plugins: [
-    ['import', {
-      'libraryName': 'ant-design-vue',
-      'libraryDirectory': 'es',
-      'style': true // `style: true` 会加载 less 文件
-    }]
+    "@vue/babel-plugin-jsx",
+    ["import", { "libraryName": "ant-design-vue", "libraryDirectory": "es", "style": "css" }] // `style: true` 会加载 less 文件
   ]
 }
