@@ -1,5 +1,17 @@
 declare module '*.vue' {
-  import Vue from 'vue';
-
-  export default Vue;
+  import { defineComponent } from 'vue'
+  const component: ReturnType<typeof defineComponent>;
+  export default component
 }
+declare module "*.less" {
+  const style: {[key: string]: string};
+  export default style;
+}
+
+
+
+// declare module '@vue/runtime-core' {
+//   interface ComponentCustomProperties {
+//     $store: Store
+//   }
+// }
