@@ -1,10 +1,11 @@
 import { ReturnGetters, Store, ActionContext } from '@/types/store';
-import Vue from "vue";
+import { ws } from "@/utils/localstorage/index";
+
 export const FOLDER_STATUS_INIT = 'FOLDER_STATUS_INIT'
 export const FOLDER_STATUS_UPDATE = 'FOLDER_STATUS_UPDATE'
 /** @typedef {typeof state} FolderState */
 const state = {
-  folderStatus:{}
+  folderStatus:{} as {x?:number,y?:number,w?:number,h?:number}
 }
 type State=typeof state
 /** @type MutationTree<FolderState> */
@@ -20,16 +21,15 @@ const mutations = {
       status.x=0
     }
     Object.assign(state.folderStatus,status)
-    Vue.ls.set('folderStatus',state.folderStatus)
+    ws().set('folderStatus',state.folderStatus)
   }
 }
 
 
 const actions = {
   initFolderStatus({commit}: ActionContext<State, Getters>){
-    let lsStatus=Vue.ls.get('folderStatus')
+    let lsStatus=ws().get('folderStatus')
     let status=Object.assign({x:0,y:0,w:200,h:200},lsStatus)
-
     commit(FOLDER_STATUS_INIT,status)
   },
   updateFolderStatus({commit}:ActionContext<State, Getters>,data :any){
@@ -39,7 +39,7 @@ const actions = {
 
 
 const getters = {
-  getFolderStatus(state: State, getters: any, rootState: Store['state'], rootGetters: any){
+  getFolderStatus(state: State){
     return state.folderStatus
   }
 }
