@@ -1,17 +1,17 @@
 <template>
   <div style="display:flex">
-    <v-left></v-left>
-    <v-right></v-right>
+
   </div>
 </template>
 
-<script>
-import Vleft from "./test1.vue";
-import Vright from "./test2.vue";
-import components from "@/components";
-export default {
-  components: components
-}
+<script lang="ts">
+import { defineComponent } from "vue";
+export default defineComponent({
+  mounted(){
+    
+  }
+  
+})
 </script>
 
 <style lang="less" scoped>
